move flattenMessages to own file to avoid error from gatsby about APIs
