+++ src/gatsby-node.js
@@ -5,23 +5,7 @@ const {
   getLanguages,
   getLanguageOption,
 } = require("./route-prefix")
-
-function flattenMessages(nestedMessages, prefix = "") {
-  return Object.keys(nestedMessages).reduce((messages, key) => {
-    let value = nestedMessages[key]
-    let prefixedKey = prefix ? `${prefix}.${key}` : key
-
-    if (typeof value === "string") {
-      messages[prefixedKey] = value
-    } else {
-      Object.assign(messages, flattenMessages(value, prefixedKey))
-    }
-
-    return messages
-  }, {})
-}
-
-exports.flattenMessages = flattenMessages
+const { flattenMessages } = require("./flattenMessages")
 
 exports.onCreateWebpackConfig = ({ actions, plugins }, pluginOptions) => {
   const {
